## v0.2.5 更新内容
### 功能与交互

- 将“下载先行卡”的执行提示从对话框改为 toast 通知，更符合当前界面的操作风格。
- 调整卡组中间区域的鼠标操作：左键仅用于拖拽卡片，右键点击卡片才会删除。
- 增加“规则上视为同名卡”的卡组数量审核。异画卡与规则同名卡会共享同一个卡位，例如“白龙之落胤”和“阿不思的落胤”合计最多投入 3 张。
- 禁限角标与投入数量审核现在会使用规则同名卡的规范 ID，并保留对原始卡片 ID 的兼容匹配。

### 性能优化

- 优化搜索结果大量出现时的渲染流程，避免一次性创建过多结果行导致界面卡顿。
- 优化搜索结果图片加载队列，移除从队列头部反复删除和逐张扫描列表的低效逻辑。
- 将搜索结果缩略图的磁盘缓存读取延后到图片加载队列中分批执行，减少搜索结果创建时的主线程阻塞。
- 在线搜索、离线搜索和先行卡搜索统一限制最多显示 500 条结果，结果过多时通过 toast 提示缩小搜索范围。
- 在线搜索请求加入过期结果检查，快速连续搜索时旧响应不会再继续写入结果列表。
- 优化卡组区域拖拽体验，缓存拖拽图标，并在拖拽期间跳过不必要的 hover 预览加载。

### 修复与维护

- 替换已弃用的 GTK/GDK 图片 API，减少新版 GTK 下的弃用警告。
- 修复导入 URL 解码中的类型警告。
- 修复部分卡片移动、整理、打乱、清空后 hover 预览状态未及时刷新的问题。
- 修复未安装 GNOME 桌面环境时部分图标缺失的问题。

+++ src/deck_url.c
@@ -242,7 +242,7 @@ bool deck_decode_from_url(
     
     // 提取d参数值（到&或字符串结尾）
     const char *param_end = strchr(d_param, '&');
-    int param_len = param_end ? (param_end - d_param) : strlen(d_param);
+    gsize param_len = param_end ? (gsize)(param_end - d_param) : strlen(d_param);
     char *base64_data = g_strndup(d_param, param_len);
     
     // 解码Base64Url为位数组
